Rename the 'key-bindings' component to 'bindkey'
I was just trying to start this name to get <Tab> to complete the name,
when I had to stop and try to remember what I'd called it.  That seems
easier to do if it's the command used for all the bindings.  Possibly a
bit less semantic, but who's counting?