Remove .zshenv
Note that dot -- this is a hidden file in a directory where the user's
home/.zshenv is supposed to be linked to the non-hidden zshenv in this
dir.

The contents duplicate everything in the real zshenv and the file isn't
even sourced on shell startup.  Its existence is a mistake. (Damn,
that's cold.)